feat: add solution for checking if a number is ugly
Added a solution for the LeetCode problem '263. Ugly Number'. The solution includes a method 'isUgly' which checks if a given number is an ugly number by dividing it by 2, 3, and 5 until it is no longer divisible by these primes. If the resulting number is 1, it returns True; otherwise, it returns False.